# Remove unused middle_nodes_table vector